// shared/types/Entities.ts
export interface BaseEntity {
  id: number
  createdAt: string
}

export enum Entities {
  GIF = 'gifs',
  CHARACTER = 'characters',
  EPISODE = 'episodes',
  SEASON = 'seasons',
  LIKE = 'likes',
  PROFILE = 'profiles',
  COMMENT = 'comments'
}

export type LikeableEntity = Entities.GIF | Entities.CHARACTER | Entities.EPISODE | Entities.SEASON